Change MyFarmFloater to pureComponent
