--- packages/webapp/src/containers/MyFarmFloater/index.js
@@ -1,24 +1,0 @@
-import Floater from "react-floater";
-import React from "react";
-import PureMyFarmFloater from "../../components/MyFarmFloater";
-import history from '../../history';
-export default function MyFarmFloater({children, openProfile}) {
-
-  const farmInfoClick = () => {
-  }
-  const farmMapClick = () => {
-}
-  const peopleClick = () => {
-  }
-
-  const Wrapper = (
-    <PureMyFarmFloater farmInfo={farmInfoClick} farmMap={farmMapClick} people={peopleClick} />
-  )
-  return (
-    <Floater component={Wrapper} placement={'bottom-end'} open={openProfile}>
-      {children}
-    </Floater>
-  )
-}
-
-
